Added formula for getopt
BSD getopt on OSX - as well as the bash builtin getopts - does not have the same arguments etc. as the Linux utils version of getopt, which is built here without gettext. Taken from rybesh's fork of homebrew ( http://github.com/rybesh/homebrew ). This is not GNU getopt(3), but can do everything it can do.