Plowshare brew patches plowshare to use ggetopt
ggetopt is installed by the brew gnu-getopt, with the prefix so that
it can be easily distinguished from the BSD getopt supplied with OSX.
Plowshare needs to use this, but expects to find it at getopt. The
patch to use the GNU tools (sed and head) is therefore changed to
include this.